Cleaned up the Database class
